refactoring Source directory (no changes to code other than moving files)
