+++ slsim/Sources/SourceCatalogues/QuasarCatalog/quasar_plus_galaxies.py
