+++ slsim/Sources/SourceCatalogues/QuasarCatalog/simple_quasar.py
