+++ slsim/Sources/SourceCatalogues/SupernovaeCatalog/__init__.py
