+++ slsim/Sources/SourceCatalogues/SupernovaeCatalog/scotch_host_data.fits
