+++ slsim/Sources/SourceCatalogues/SupernovaeCatalog/supernovae_data.pkl
