+++ slsim/Sources/SourcePopulation/galaxies.py
@@ -2,7 +2,7 @@ import numpy as np
 import numpy.random as random
 from slsim.selection import object_cut
 from slsim.Util import param_util
-from slsim.Sources.source_pop_base import SourcePopBase
+from slsim.Sources.SourcePopulation.source_pop_base import SourcePopBase
 from astropy.table import Column, vstack
 from slsim.Util.param_util import (
     average_angular_size,
