+++ slsim/Sources/__init__.py
@@ -1,8 +1,7 @@
-from .galaxies import Galaxies
-from .point_sources import PointSources
-from .point_plus_extended_sources import PointPlusExtendedSources
-from . import QuasarCatalog
-from . import SupernovaeCatalog
+from slsim.Sources.SourcePopulation.galaxies import Galaxies
+from slsim.Sources.SourcePopulation.point_sources import PointSources
+from slsim.Sources.SourcePopulation.point_plus_extended_sources import PointPlusExtendedSources
+from .SourceCatalogues import QuasarCatalog, SupernovaeCatalog
 
 __all__ = [
     "Galaxies",
