+++ slsim/lens_pop.py
@@ -3,7 +3,7 @@ import numpy as np
 from slsim.lens import Lens
 from typing import Optional
 from astropy.cosmology import Cosmology
-from slsim.Sources.source_pop_base import SourcePopBase
+from slsim.Sources.SourcePopulation.source_pop_base import SourcePopBase
 from slsim.LOS.los_pop import LOSPop
 from slsim.Deflectors.DeflectorPopulation.deflectors_base import DeflectorsBase
 from slsim.lensed_population_base import LensedPopulationBase
