+++ tests/test_Sources/test_SourcePopulation/test_quasars.py
@@ -1,7 +1,7 @@
 from astropy.cosmology import FlatLambdaCDM
 from astropy.units import Quantity
-from slsim.Sources.point_sources import PointSources
-from slsim.Sources.QuasarCatalog.simple_quasar import quasar_catalog_simple
+from slsim.Sources.SourcePopulation.point_sources import PointSources
+from slsim.Sources.SourceCatalogues.QuasarCatalog.simple_quasar import quasar_catalog_simple
 import pytest
