+++ tests/test_Sources/test_SourceTypes/test_interpolated_image.py
