+++ tests/test_Sources/test_SourceVariability/test_light_curve_interpolation.py
