+++ tests/test_Sources/test_SourceVariability/test_sinusoidal.py
