+++ tests/test_Sources/test_Supernovae/test_supernovae_host_match.py
@@ -2,8 +2,8 @@ import pytest
 import numpy as np
 from astropy.units import Quantity
 from astropy.cosmology import FlatLambdaCDM
-from slsim.Sources.galaxy_catalog import GalaxyCatalog
-from slsim.Sources.supernovae_host_match import SupernovaeHostMatch
+from slsim.Sources.SourceCatalogues.skypy_galaxy_catalog import GalaxyCatalog
+from slsim.Sources.Supernovae.supernovae_host_match import SupernovaeHostMatch
 
 
 skypy_config = None
